requirements.txt
# FastAPI学习项目依赖文件
# 这个文件定义了项目运行所需的所有Python包

# FastAPI - 现代、快速的Web框架，用于构建API
fastapi==0.104.1

# Uvicorn - ASGI服务器，用于运行FastAPI应用
uvicorn[standard]==0.24.0

# SQLAlchemy - Python SQL工具包和对象关系映射(ORM)
sqlalchemy==1.4.53

# Pydantic - 数据验证和设置管理，使用Python类型提示
pydantic==2.5.0

# Python-multipart - 处理表单数据和文件上传
python-multipart==0.0.6

# bcrypt - 密码哈希库（直接使用，替代passlib）
bcrypt==4.1.2

# Python-jose - JSON Web Token实现
python-jose[cryptography]==3.3.0

# Python-decouple - 从环境变量和.ini文件中分离设置
python-decouple==3.8